separated invoice component
